# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Contacts app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is documentation of contacts app"
  },
  "servers": [
    {
      "url": "https://nodejs-hw-mongodb-1-3mbt.onrender.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
...
  "paths": {
    "/contacts": {
      "post": {
        "tags": ["Contacts"],
        "summary": "Create a New Contact",
        "operationId": "createContact",
        "description": "**Create a new contact** with detailed information using this endpoint.\n\n### Required Fields\n- **Name**: The full name of the contact.\n- **Phone Number**: The contact's phone number, including country code if necessary.\n- **Contact Type**: The type of contact (e.g., work, home, personal).\n\n### Optional Fields\n- **Email**: The contact's email address (optional).\n- **Is Favorite**: A boolean flag indicating whether the contact is marked as a favorite (optional).\n- **Photo**: An optional image file for the contact's photo.\n\nEnsure that all required fields are provided to successfully create a contact. You may also include additional optional details to enrich the contact's information.\n\n### Security\n- This endpoint requires authentication via a Bearer Token. Include a valid token in the `Authorization` header.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["name", "phoneNumber", "contactType"],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "The full name of the contact.",
                    "example": "Nik Brook"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "description": "The contact's phone number, including country code if necessary.",
                    "example": "+38000000899"
                  },
                  "email": {
                    "type": "string",
                    "description": "The contact's email address. This is optional.",
                    "example": "nik@example.com"
                  },
                  "isFavourite": {
                    "type": "boolean",
                    "description": "Indicates whether the contact is marked as a favorite.",
                    "example": false
                  },
                  "contactType": {
                    "type": "string",
                    "description": "The type of the contact. It can be 'work', 'home', or 'personal'.",
                    "enum": ["work", "home", "personal"],
                    "example": "personal"
                  },
                  "photo": {
                    "type": "string",
                    "format": "binary",
                    "description": "An optional file for the contact's photo. It should be in image format.\n**If left empty, this field should not be included in the request.**\n"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "**Contact Successfully Created**\nThe contact was created successfully. The response contains details of the newly created contact.\n",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201,
                      "description": "HTTP status code"
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully created a contact!",
                      "description": "Confirmation message indicating successful creation"
                    },
                    "data": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/contact"
                        },
                        {
                          "type": "object",
                          "properties": {
                            "_id": {
                              "type": "string",
                              "example": "66cb5f4d0318528063e77c98",
                              "description": "Unique identifier for the created contact"
                            },
                            "createdAt": {
                              "type": "string",
                              "format": "date-time",
                              "example": "2024-08-29T20:39:40.775Z",
                              "description": "The date and time when the contact was created"
                            },
                            "updatedAt": {
                              "type": "string",
                              "format": "date-time",
                              "example": "2024-08-29T20:39:40.775Z",
                              "description": "The date and time when the contact was last updated"
                            }
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid or missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 400,
                      "description": "HTTP status code"
                    },
                    "message": {
                      "type": "string",
                      "example": "BadRequestError",
                      "description": "Error message indicating the nature of the bad request"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Bad request, body parameters are incorrect",
                          "description": "Detailed error message"
                        },
                        "errors": {
                          "type": "string",
                          "example": "'name' is a required field; 'phoneNumber' is a required field; 'contactType' is a required field",
                          "description": "List of validation errors"
                        }
                      }
                    }
                  },
                  "required": ["status", "message", "data"]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "description": "InternalServerError",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "InternalServerError"
                    },
                    "data": {
                      "type": "object",
                      "required": ["message"],
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Failed to save photo, please try again later."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Contacts"],
        "summary": "Retrieve a List of Contacts",
        "operationId": "getContacts",
        "description": "**Retrieve a list of contacts** with advanced filtering, sorting, and pagination options.\n\nThis endpoint allows you to:\n- **Filter** contacts by various attributes such as name, phone number, email, and more.\n- **Sort** contacts based on any field in either ascending or descending order.\n- **Paginate** results to efficiently manage large datasets.\n\n### Security\n- Requires a valid Bearer Token to be provided in the `Authorization` header.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "example": 1,
            "description": "**Page Number**\nThe page number to retrieve. Defaults to `1` if not specified.\n"
          },
          {
            "in": "query",
            "name": "perPage",
            "schema": {
              "type": "integer"
            },
            "example": 10,
            "description": "**Contacts per Page**\nThe number of contacts to display per page. Defaults to `10` if not specified.\n"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string"
            },
            "example": "createdAt",
            "description": "**Sorting Field**\nThe field by which to sort the contacts. Any contact field can be used.\nDefaults to `_id` if not specified.\n"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"]
            },
            "example": "asc",
            "description": "**Sorting Order**\nThe order to sort by, either ascending (`asc`) or descending (`desc`).\n"
          },
          {
            "in": "query",
            "name": "isFavourite",
            "schema": {
              "type": "boolean"
            },
            "example": true,
            "description": "**Filter by Favorite Status**\nFilter contacts that are marked as favorites (`true`) or not (`false`).\n"
          },
          {
            "in": "query",
            "name": "contactType",
            "schema": {
              "type": "string",
              "enum": ["work", "home", "personal"]
            },
            "example": "personal",
            "description": "**Filter by Contact Type**\nFilter contacts by type (e.g., `work`, `home`, `personal`). Defaults to `personal` if not specified.\n"
          }
        ],
        "responses": {
          "200": {
            "description": "**Contacts Retrieved Successfully**\nThe contact list was successfully retrieved. The response includes details of the contacts, pagination information, and other relevant metadata.\n",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200,
                      "description": "HTTP status code"
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully retrieved contacts!",
                      "description": "Confirmation message indicating successful retrieval"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "data",
                        "page",
                        "perPage",
                        "totalItems",
                        "totalPages",
                        "hasNextPage",
                        "hasPreviousPage"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "allOf": [
                              {
                                "$ref": "#/components/schemas/contact"
                              },
                              {
                                "type": "object",
                                "properties": {
                                  "_id": {
                                    "type": "string",
                                    "example": "66cb5f4d0318528063e77c98",
                                    "description": "Unique identifier for the contact"
                                  },
                                  "createdAt": {
                                    "type": "string",
                                    "format": "date-time",
                                    "example": "2024-08-29T20:39:40.775Z",
                                    "description": "The date and time when the contact was created"
                                  },
                                  "updatedAt": {
                                    "type": "string",
                                    "format": "date-time",
                                    "example": "2024-08-29T20:39:40.775Z",
                                    "description": "The date and time when the contact was last updated"
                                  }
                                }
                              }
                            ]
                          }
                        },
                        "page": {
                          "type": "integer",
                          "example": 1,
                          "description": "Current page number"
                        },
                        "perPage": {
                          "type": "integer",
                          "example": 10,
                          "description": "Number of contacts per page"
                        },
                        "totalItems": {
                          "type": "integer",
                          "example": 4,
                          "description": "Total number of contacts matching the query"
                        },
                        "totalPages": {
                          "type": "integer",
                          "example": 1,
                          "description": "Total number of pages available"
                        },
                        "hasNextPage": {
                          "type": "boolean",
                          "example": false,
                          "description": "Whether there are more pages available after the current page"
                        },
                        "hasPreviousPage": {
                          "type": "boolean",
                          "example": false,
                          "description": "Whether there are pages available before the current page"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "contact": {
        "type": "object",
        "required": ["name", "phoneNumber", "contactType"],
        "properties": {
          "name": {
            "description": "The full name of the contact",
            "type": "string",
            "example": "Nik Brook"
          },
          "phoneNumber": {
            "description": "The contact's phone number",
            "type": "string",
            "example": "+38000000899"
          },
          "email": {
            "description": "The contact's email address",
            "type": "string",
            "example": "nik@example.com"
          },
          "isFavorite": {
            "description": "Indicates whether the contact is marked as a favorite",
            "type": "boolean",
            "example": false
          },
          "contactType": {
            "description": "The category of the contact (e.g., work, home, personal)",
            "type": "string",
            "enum": ["work", "home", "personal"],
            "example": "personal"
          },
          "userId": {
            "description": "The ID of the user who owns this contact",
            "type": "string",
            "example": "66cb5ed80318528063e77c87"
          },
          "photo": {
            "description": "URL to the contact's photo",
            "type": "string",
            "example": "https://res.cloudinary.com/dg5fttpzf/image/upload/v1724604273/vnytdl8f8xwylaaaud6b.jpg"
          }
        }
      }
    },
    "responses": {
...
      "401": {
        "description": "**Unauthorized**\n\nAuthentication is required or the provided token is invalid.\n\n- **message**: Detailed message about the authorization issue.\n",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["status", "message", "data"],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401,
                  "description": "HTTP status code indicating unauthorized access"
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError",
                  "description": "Error message indicating authentication issues"
                },
                "data": {
                  "type": "object",
                  "required": ["message"],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User not authorized",
                      "description": "Detailed message about the authorization issue"
                    }
                  },
                  "additionalProperties": false
                }
              }
            },
            "examples": {
              "no_token": {
                "summary": "User not authorized",
                "value": {
                  "status": 401,
                  "message": "UnauthorizedError",
                  "data": {
                    "message": "User not authorized"
                  }
                }
              },
              "token_expired": {
                "summary": "Access token expired",
                "value": {
                  "status": 401,
                  "message": "UnauthorizedError",
                  "data": {
                    "message": "Access token expired"
                  }
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Not Found - Session or Authorization header not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["status", "message", "data"],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404,
                  "description": "HTTP status code"
                },
                "message": {
                  "type": "string",
                  "example": "NotFoundError",
                  "description": "Error message indicating that the resource was not found"
                },
                "data": {
                  "type": "object",
                  "required": ["message"],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Session not found",
                      "description": "Detailed message about the missing session or header"
                    }
                  },
                  "additionalProperties": false
                }
              }
            },
            "examples": {
              "no_token": {
                "summary": "Session not found",
                "value": {
                  "status": 404,
                  "message": "NotFoundError",
                  "data": {
                    "message": "Session not found"
                  }
                }
              },
              "token_expired": {
                "summary": "Authorization header not found",
                "value": {
                  "status": 404,
                  "message": "NotFoundError",
                  "data": {
                    "message": "Authorization header not found"
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }
}